update home banner style
