+++ src/app/(main)/components/HomeBanner/HomeBanner.module.css
@@ -1,0 +1,205 @@
+
+.section_banner {
+    display: grid;
+    grid-template-columns: repeat(8, 1fr);
+    grid-template-rows: repeat(6, 1fr);
+    background-color: var(--clr-one-400);
+
+    border-bottom-left-radius: 30px;
+    border-bottom-right-radius: 30px;
+    margin-bottom: 20px;
+    aspect-ratio: 1;
+}
+.image_container {
+    grid-column: 1/-1;
+    grid-row: 1/-1;
+    max-width: 100%;
+    height: 100%;
+    object-fit: cover;
+    
+    border-bottom-left-radius: 30px;
+    border-bottom-right-radius: 30px;
+}
+.sheet{
+    grid-column: 1/-1;
+    grid-row: 1/-1;
+    overflow: hidden;
+
+    display: flex;
+    justify-content: space-evenly;
+    align-items: center;
+    flex-direction: column;
+    gap:10px;
+}
+.infoBanner{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    background-color: rgba(255, 255, 255, 0.85);
+    width: fit-content;
+    padding: 1rem 1.3rem;
+    border-radius: 10px;
+
+}
+.title{
+    font-family: var(--ff-one);
+    color: var(--clr-secondary-2);
+    font-weight: 400;
+    font-size: 1.8rem;
+    text-align: center;
+    margin-bottom: 8px;
+    max-width: 305px;
+}
+.titleAccent{
+    color: var(--clr-primary-2);
+    font-family: var(--ff-two);
+    font-style: italic;
+    font-weight: 300;
+}
+.message{
+    font-family: var(--ff-one);
+    font-size: 0.9rem;
+    text-align: center;
+    max-width: 235px;
+}
+
+.modalBtn{
+    background-color: white;
+    color: var(--clr-primary-2);
+    font-family: var(--ff-one);
+    border: solid 1.5px var(--clr-primary-2);
+    padding: 0.5rem 1.3rem;
+    border-radius: 6px;
+    font-weight: 500;
+    font-size: 1.2rem;
+}
+.modalBtn:hover{
+    background-color: var(--clr-primary-2);
+    color: white;
+    border-color: white;
+}
+
+@media (min-width: 500px) {
+    .section_banner {
+        grid-template-columns: repeat(12, 1fr);
+        grid-template-rows: repeat(8, 1fr);
+
+        border-radius: 10px;
+        margin: 1rem;
+        aspect-ratio: unset;
+    }
+
+    .image_container {
+        grid-column: 1/6;
+        grid-row: 1/-1;
+        max-width: 100%;
+
+        border-radius: 0px;
+    }
+    .sheet{
+        grid-column: 6/-1;
+        grid-row: 1/-1;
+    }
+
+    .infoBanner{
+        padding: 0.8rem 1rem;
+    }
+    .title{
+        font-size: 1.2rem;
+        margin-bottom: 8px;
+        max-width: 305px;
+    }
+    .message{
+        font-size: 0.7rem;
+        text-align: left;
+        max-width: 220px;
+    }
+}
+
+@media (min-width: 992px) {
+    .section_banner {
+        max-height: 411.6px;
+    }
+    .infoBanner{
+        padding: 0.8rem 1rem;
+    }
+    .title{
+        font-size: 1.2rem;
+        margin-bottom: 8px;
+        max-width: 305px;
+    }
+    .message{
+        font-size: 0.7rem;
+        text-align: left;
+        max-width: 220px;
+        margin-bottom: 1rem;
+    }
+    .modalBtn{
+        padding: 0.7rem 1.6rem;
+        border-radius: 6px;
+        font-size: 1.4rem;
+    }
+}
+
+@media (min-width: 670px) {
+    .title{
+        font-size: clamp(1.8rem, 4vw, 3.2rem);
+        margin-bottom: 8px;
+        max-width: clamp(305px, 40vw, 440px);
+    }
+    .message{
+        font-size: clamp(0.8rem, 1.6vw, 1.4rem);
+        text-align: left;
+        max-width: clamp(273px, 35vw, 420px);
+    }
+}
+@media (min-width: 1200px) {
+
+    /* ------------------- Banner Section */
+
+    /* .text_banner {
+        padding-left: 30px;
+        margin-top: 20px;
+        text-align: left;
+    }
+
+    .title_uniendo {
+        font: var(--t-desktop-four);
+        letter-spacing: 4px;
+    }
+
+    .title_cuidadores {
+        font: var(--t-desktop-five);
+        letter-spacing: 4px;
+    }
+
+    .detail_banner {
+        margin-left: 0px;
+    } */
+
+}
+
+@media (min-width: 1290px) {
+
+    /* ------------------- Banner Section */
+
+    /* .title_uniendo {
+        font: var(--t-desktop-one);
+    }
+
+    .title_cuidadores {
+        font: var(--t-desktop-two);
+    }
+
+    .detail_banner {
+        font: var(--tx-desktop-four);
+    }
+
+    #modal_button {
+        font: var(--tx-desktop-two);
+    }
+
+    #modal_button:hover {
+        font: var(--tx-desktop-two);
+    } */
+}
