+++ src/app/(main)/page.jsx
@@ -1,44 +1,14 @@
-'use client';
-import { useState } from 'react';
 import Link from 'next/link';
 import styles from './page.module.css';
 import { IoIosArrowForward } from "react-icons/io";
-import Modal from '../(auth)/auth/login/ModalLogin/ModalLogin';
+import HomeBanner from './components/HomeBanner/HomeBanner';
 
 export default function Home() {
-  const [isOpen, setIsOpen] = useState(false);
-  const openModal = () => setIsOpen(true);
-  const closeModal = () => setIsOpen(false);
 
   return (
     <main className={styles.main}>
       <div className={styles.section}>
-
-        {/* ----- Seccion Banner ----- */}
-
-        <section className={styles.section_banner}>
-          <img className={styles.image_container} src="https://firebasestorage.googleapis.com/v0/b/c16-ronda.appspot.com/o/imagenes%2Fmanos_unidas.png?alt=media&token=f47a7fc7-7178-4e91-932f-11e43f68d4d"></img>
-
-          <div className={styles.text_container}>
-            <div className={styles.text_banner}>
-              <h1>
-                <span className={styles.title_uniendo}>Uniendo</span>
-                <span className={styles.title_uniendo}> a</span>
-                <span className={styles.title_cuidadores}> Cuidadores</span>
-              </h1>
-              <p className={styles.detail_banner}>
-                Únete a la red más grande de apoyo para cuidadores
-              </p>
-            </div>
-              <button onClick={() => setIsOpen(true)} id={styles.modal_button}>
-                {' '}
-                Inscríbete
-                {' '}
-              </button>
-            <Modal isOpen={isOpen} onClose={closeModal} />
-          </div>
-        </section>
-
+        <HomeBanner />
         {/* ----- Welcome Section ----- */}
 
         <section className={styles.welcome_section}>
