+++ src/app/(main)/page.module.css
@@ -12,90 +12,6 @@
   grid-column: 1/-1;
 }
 
-/* ------------------- Banner Section */
-
-.section_banner {
-  display: grid;
-  grid-template-columns: repeat(8, 1fr);
-  grid-template-rows: repeat(6, 1fr);
-  background-color: var(--clr-one-400);
-
-  border-bottom-left-radius: 30px;
-  border-bottom-right-radius: 30px;
-  margin-bottom: 20px;
-}
-
-.image_container {
-  grid-column: 1/-1;
-  grid-row: 1/-1;
-  max-width: 100%;
-
-  border-bottom-left-radius: 30px;
-  border-bottom-right-radius: 30px;
-}
-
-.text_container {
-  grid-column: 2/-2;
-  grid-row: 2/-2;
-  align-items: center;
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  grid-template-rows: repeat(3, 1fr);
-
-}
-
-.text_banner {
-  background-color: rgba(255, 255, 255, 0.8);
-  border-radius: 10px;
-  align-items: center;
-  grid-column: 1/-1;
-  grid-row: 1/3;
-  padding-top: 20px;
-}
-
-.title_uniendo {
-  font: var(--t-mobile-four);
-}
-
-.title_cuidadores {
-  font: var(--t-mobile-two);
-  color: var(--clr-primary-2);
-  font-weight: bolder;
-}
-
-.detail_banner {
-  text-align: center;
-  font: var(--tx-mobile-one);
-  margin: 20px;
-}
-
-#modal_button {
-  font: var(--tx-mobile-three);
-  color: var(--clr-primary-2);
-  background-color: rgba(255, 255, 255, 1);
-  border-radius: 10px;
-  border: 1px solid var(--clr-primary-2);
-  cursor: pointer;
-  grid-column: 2/-2;
-  grid-row: 3/-1;
-  margin-bottom: 0;
-  padding-top: 10px;
-  padding-bottom: 10px;
-}
-
-#modal_button:hover {
-  font: var(--tx-mobile-three);
-  color: white;
-  background-color: var(--clr-primary-2);
-  border-radius: 10px;
-  border: 1px solid var(--clr-primary-2);
-  cursor: pointer;
-  grid-column: 2/-2;
-  grid-row: 3/-1;
-  padding-top: 10px;
-  padding-bottom: 10px;
-}
-
 /* ------------------- Welcome Section */
 
 .welcome_title {
@@ -290,61 +206,6 @@
     margin: 30px;
   }
 
-  /* ------------------- Banner Section */
-
-  .section_banner {
-    grid-template-columns: repeat(12, 1fr);
-    grid-template-rows: repeat(8, 1fr);
-
-    border-radius: 30px;
-    margin-top: 20px;
-  }
-
-  .image_container {
-    grid-column: 1/6;
-    grid-row: 1/-1;
-    max-width: 100%;
-
-    border-top-left-radius: 30px;
-    border-bottom-left-radius: 30px;
-    border-bottom-right-radius: 0px;
-  }
-
-  .text_container {
-    grid-column: 7/-2;
-    grid-row: 1/-1;
-  }
-
-  .text_banner {
-    border-radius: 30px;
-  }
-
-  .title_uniendo {
-    font: var(--t-mobile-four);
-  }
-
-  .title_cuidadores {
-    font: var(--t-mobile-two);
-  }
-
-  .detail_banner {
-    text-align: left;
-    font: var(--tx-mobile-one);
-    margin: 30px;
-  }
-
-  #modal_button {
-    padding-top: 15px;
-    padding-bottom: 15px;
-    border-radius: 20px;
-  }
-
-  #modal_button:hover {
-    padding-top: 15px;
-    padding-bottom: 15px;
-    border-radius: 20px;
-  }
-
   /* ------------------- Welcome Section */
   .welcome_title {
     font: var(--t-desktop-five);
@@ -358,22 +219,6 @@
   .main {
     max-width: 1290px;
     margin: 0 auto;
-  }
-
-  /* ------------------- Banner Section */
-  
-  .title_uniendo {
-    font: var(--t-mobile-six);
-    letter-spacing: 2px;
-  } 
-
-  .title_cuidadores {
-    font: var(--t-mobile-five);
-    letter-spacing: 2px;
-  }
-
-  .detail_banner {
-    font: var(--tx-desktop-four);
   }
 
   /* ------------------- Welcome Section */
@@ -418,57 +263,6 @@
 }
 
 /* ------------------- Pantalla Extra-Grande */
-
-
-@media (min-width: 1200px) {
-
-/* ------------------- Banner Section */
-
-  .text_banner {
-    padding-left: 30px;
-    margin-top: 20px;
-    text-align: left;
-  }
-
-  .title_uniendo {
-    font: var(--t-desktop-four);
-    letter-spacing: 4px;
-  }
-
-  .title_cuidadores {
-    font: var(--t-desktop-five);
-    letter-spacing: 4px;
-  }
-
-  .detail_banner {
-    margin-left: 0px;
-  }
-
-}
-@media (min-width: 1290px) {
-
-/* ------------------- Banner Section */
-  
-  .title_uniendo {
-    font: var(--t-desktop-one);
-  }
-
-  .title_cuidadores {
-    font: var(--t-desktop-two);
-  }
-
-  .detail_banner {
-    font: var(--tx-desktop-four);
-  }
-
-  #modal_button {
-    font: var(--tx-desktop-two);
-  }
-
-  #modal_button:hover {
-    font: var(--tx-desktop-two);
-  }
-}
 
 @media (min-width: 1400px) {
